lots of java updates

+++ java.m4
@@ -1,7 +1,7 @@
 AC_DEFUN([ONMS_CHECK_JDK],
   [
     AC_ARG_WITH([java],
-      [AS_HELP_STRING([--with-java=JAVA_HOME], [set the path to JAVA_HOME for the jdk])],
+      [AS_HELP_STRING([--with-java=JAVA_HOME], [set the path to JAVA_HOME for the JDK])],
       [],
       [with_java=check])
 
@@ -10,17 +10,17 @@ AC_DEFUN([ONMS_CHECK_JDK],
       [],
       [with_jvm_arch=none])
 
-    AS_IF([test "x$with_java" = "xno"], [AC_MSG_ERROR([the path to a jdk is required to build jrrd])])
+    AS_IF([test "x$with_java" = "xno"], [AC_MSG_ERROR([the path to a JDK is required to build jrrd])])
     AS_IF([test "x$with_java" = "xyes"], [AC_MSG_ERROR([the argument to --with-java must specify a JDK])])
     AS_IF([test "x$with_java" = "xcheck"], 
-          [ONMS_FIND_JDK($1, ["$with_jvm_arch"])],
-          [ONMS_VALIDATE_JDK(["$with_java"], $1, ["$with_jvm_arch"])]
+          [ONMS_FIND_JDK([$1], [$with_jvm_arch])],
+          [ONMS_VALIDATE_JDK([$with_java], [$1], [$with_jvm_arch])]
     )
 
     AS_IF([test "x$HAS_JDK" = "x" || test "$HAS_JDK" = "false" || test "$HAS_JDK" = "no"],
-          [AC_MSG_ERROR([unable to find a valid jdk for java version $1])])
+          [AC_MSG_ERROR([unable to find a valid JDK for java version $1])])
 
-    AC_MSG_NOTICE([using jdk at $JAVA_HOME])
+    AC_MSG_NOTICE([using JDK at $JAVA_HOME])
 
     case $host_os in
         darwin*)
@@ -45,7 +45,7 @@ AC_DEFUN([ONMS_CHECK_JDK],
 
 AC_DEFUN([ONMS_FIND_JDK],
   [
-    AC_MSG_NOTICE([searching for a $1 jdk])
+    AC_MSG_NOTICE([searching for a $1 JDK])
     
     HAS_JDK=no
     AS_IF([test "x$JAVA_HOME" != "x"], 
@@ -60,9 +60,9 @@ AC_DEFUN([ONMS_FIND_JDK],
          do
              java_from_path=`readlink $java_from_path`
          done
-         java_home_from_path=`AS_DIRNAME(["$java_from_path"])`
-         java_home_from_path=`AS_DIRNAME(["$java_home_from_path"])`
-	 _ONMS_TRY_JAVA_DIR([$java_home_from_path], [$1], [AC_MSG_NOTICE([attempting to find the jdk for $java_from_path.])], [$2])
+         java_home_from_path=`AS_DIRNAME("$java_from_path")`
+         java_home_from_path=`AS_DIRNAME("$java_home_from_path")`
+         _ONMS_TRY_JAVA_DIR([$java_home_from_path], [$1], [AC_MSG_NOTICE([attempting to find the JDK for $java_from_path.])], [$2])
          AS_UNSET([java_from_path])
          AS_UNSET([java_home_from_path])
       ]
@@ -95,22 +95,37 @@ AC_DEFUN([_ONMS_TRY_JAVA_DIR],
 
 AC_DEFUN([ONMS_VALIDATE_JDK],
   [
-    AC_MSG_NOTICE([checking if $1 is home for a valid $2 jdk])
+    AC_MSG_NOTICE([checking if $1 is home for a valid $2 JDK])
+
+    AC_ARG_ENABLE(jdk-validation,
+      [  --disable-jdk-validation       don't validate the JDK
+        --enable-jdk-validation        make sure the JDK is valid])
 
     HAS_JDK=yes
-    dnl the following so HAS_JDK of they fail to pass the check
-    _ONMS_CHECK_FOR_JAVA($1)
-    _ONMS_CHECK_FOR_JAVAC($1)
-    _ONMS_CHECK_FOR_JAR($1)   
-    _ONMS_CHECK_FOR_JAVAH($1)
-    _ONMS_CHECK_JAVA_VERSION($2)
-    _ONMS_CHECK_JAVA_ARCH($3)
-    _ONMS_CHECK_FOR_JNI_HEADERS($1)
 
-    AS_IF([test "$HAS_JDK" != yes], 
-          [AC_MSG_NOTICE([no valid jdk found at $1])],
-          [AC_MSG_NOTICE([found a valid jdk. setting JAVA_HOME to $1]); JAVA_HOME="$1"]
-    )
+    if test "x$enable_jdk_validation" != xno; then
+
+      dnl the following so HAS_JDK of they fail to pass the check
+      _ONMS_CHECK_FOR_JAVA($1)
+      _ONMS_CHECK_FOR_JAVAC($1)
+      _ONMS_CHECK_FOR_JAR($1)   
+      _ONMS_CHECK_FOR_JAVAH($1)
+      _ONMS_CHECK_JAVA_VERSION($2)
+      _ONMS_CHECK_JAVA_ARCH($3)
+      _ONMS_CHECK_FOR_JNI_HEADERS($1)
+
+      AS_IF([test "$HAS_JDK" != yes], 
+            [AC_MSG_NOTICE([no valid JDK found at $1])],
+            [AC_MSG_NOTICE([found a valid JDK. setting JAVA_HOME to $1]); JAVA_HOME="$1"]
+      )
+
+    else
+
+      AC_MSG_NOTICE([JDK validation was skipped])
+      _ONMS_CHECK_JAVA_ARCH($3)
+      _ONMS_CHECK_FOR_JNI_HEADERS($1)
+
+    fi
 
   ]
 )
@@ -121,13 +136,12 @@ AC_DEFUN([_ONMS_CHECK_FOR_JNI_HEADERS],
       [
         AC_MSG_CHECKING([for jni headers])
         HAS_JNI_HEADERS=yes
-	AS_IF([test -d "$1/include" && test -f "$1/include/jni.h"],
+        AS_IF([test -d "$1/include" && test -f "$1/include/jni.h"],
           [
-            JNI_INCLUDES=`find "$1/include/" -type d | while read dir
-            do
-	      echo $ECHO_N "-I$dir "
+            JNI_INCLUDES=`find "$1/include" -type d | while read DIR; do
+               printf -- "-I\$DIR "
             done`
-	  ],
+          ],
           [ dnl no include directory so invalid jdk
             HAS_JNI_HEADERS=no
             HAS_JDK=no
@@ -170,31 +184,36 @@ AC_DEFUN([_ONMS_CHECK_JAVA_ARCH],
       [
         HAS_VALID_JAVA_ARCH=yes
         AC_MSG_CHECKING([if java architecture meets requirements])
-        _ONMS_CREATE_JAVA_SRC([getarch], [System.out.println(System.getProperty("sun.arch.data.model", "32"));])
-        _ONMS_COMPILE_SOURCE_FILE([getarch.java], [tmp-classes], [])
-        JAVA_ARCH=`"$JAVA" -cp tmp-classes getarch`
-        rm -rf tmp-classes
-        rm -f getarch.java
 
-        AS_IF([test "x$1" != "xnone" && test "$JAVA_ARCH" != "$1"],
-          [
-            AC_MSG_CHECKING([if java architecture meets requirements with -d$1])
-            _ONMS_CREATE_JAVA_SRC([getarch], [System.out.println(System.getProperty("sun.arch.data.model", "32"));])
-            _ONMS_COMPILE_SOURCE_FILE([getarch.java], [tmp-classes], [])
-            JAVA_ARCH=`"$JAVA" -d$1 -cp tmp-classes getarch`
-            rm -rf tmp-classes
-            rm -f getarch.java
-
-            AS_IF([test "x$1" != "xnone" && test "$JAVA_ARCH" != "$1"],
-              [
-                HAS_VALID_JAVA_ARCH=no
-                HAS_JDK=no
-              ]
-            )
-          ]
-        )
+        if test "x$with_jvm_arch" != "xnone"; then
+          JAVA_ARCH="$with_jvm_arch"
+        else
+          _ONMS_CREATE_JAVA_SRC([getarch], [System.out.println(System.getProperty("sun.arch.data.model", "32"));])
+          _ONMS_COMPILE_SOURCE_FILE([getarch.java], [tmp-classes], [])
+          JAVA_ARCH=`"$JAVA" -cp tmp-classes getarch`
+          rm -rf tmp-classes
+          rm -f getarch.java
+  
+          AS_IF([test "x$1" != "xnone" && test "$JAVA_ARCH" != "$1"],
+            [
+              AC_MSG_CHECKING([if java architecture meets requirements with -d$1])
+              _ONMS_CREATE_JAVA_SRC([getarch], [System.out.println(System.getProperty("sun.arch.data.model", "32"));])
+              _ONMS_COMPILE_SOURCE_FILE([getarch.java], [tmp-classes], [])
+              JAVA_ARCH=`"$JAVA" -d$1 -cp tmp-classes getarch`
+              rm -rf tmp-classes
+              rm -f getarch.java
+  
+              AS_IF([test "x$1" != "xnone" && test "$JAVA_ARCH" != "$1"],
+                [
+                  HAS_VALID_JAVA_ARCH=no
+                  HAS_JDK=no
+                ]
+              )
+            ]
+          )
+        fi
         AC_MSG_RESULT([$HAS_VALID_JAVA_ARCH, $JAVA_ARCH-bit])
-	AC_SUBST(JAVA_ARCH)
+        AC_SUBST(JAVA_ARCH)
       ]
     )
   ]
